workin on displaying posts with added images
