+++ js/post_feed.mjs
@@ -1,6 +1,6 @@
-import { posts, authorization } from "./API_URLs.mjs";
+import { posts, authorization, testSomething } from "./API_URLs.mjs";
 
-const displaySomething = document.querySelector(".col-sm-7");    
+    
     
 async function getPosts(url) {
     try{
@@ -15,17 +15,7 @@ async function getPosts(url) {
         const response = await fetch(url, getData);
         console.log(response);
         const json = await response.json();
-        for (let i = 0; i < json.length; i++){
-            console.log(json[i]);
-            displaySomething.innerHTML += `<div class="card mb-3">
-            <img src="${json[i].media}" class="card-img-top" alt="black and white picture of a huge cliff by the ocean">
-            <div class="card-body">
-                <h5 class="card-title">${json[i].author.name}</h5>
-                <p class="card-text"><small class="text-muted">Published ${json[i].updated}</small></p>
-            </div>
-        </div>`;
-
-        }
+        testSomething(json);
        
     } catch (error) {
        console.log(error);
